add oauth2

+++ src/main/java/com/javamaster/repository/UserEntityRepository.java
@@ -3,7 +3,8 @@ package com.javamaster.repository;
 import com.javamaster.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserEntityRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
 
     UserEntity findByLogin(String login);
+
 }
